# `chacha20`: Improve AVX2 performance