- A bunch of instructions for accessing the 128-bit lanes have been replaced by a union.
- I've implemented a widely-used ChaCha optimisation (that I spotted in `c2-chacha`).